@@ -66,105 +71,103 @@
     ]
   },
   "dependencies": {
-    "@graphql-modules/logger": "^0.2.0",
-    "@nrwl/nx": "7.0.2",
-    "@types/strip-json-comments": "^0.0.30",
-    "karma-spec-reporter": "^0.0.32",
-    "karma-verbose-reporter": "^0.0.6",
-    "request": "2.88.0"
+    "@nrwl/nx": "7.1.1",
+    "resize-observer-polyfill": "^1.5.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.10.6",
-    "@angular-devkit/build-ng-packagr": "0.10.6",
-    "@angular-devkit/core": "0.8.7",
-    "@angular-devkit/schematics": "0.8.7",
-    "@angular/animations": "7.1.0",
-    "@angular/cdk": "7.1.0",
-    "@angular/cli": "7.0.6",
-    "@angular/common": "7.1.0",
-    "@angular/compiler": "7.1.0",
-    "@angular/compiler-cli": "7.1.0",
-    "@angular/core": "7.1.0",
-    "@angular/flex-layout": "7.0.0-beta.19",
-    "@angular/forms": "7.1.0",
-    "@angular/language-service": "7.1.0",
-    "@angular/material": "7.1.0",
-    "@angular/platform-browser": "7.1.0",
-    "@angular/platform-browser-dynamic": "7.1.0",
-    "@angular/router": "7.1.0",
-    "@nrwl/angular-console-enterprise-electron": "^0.0.34",
-    "@nrwl/angular-console-enterprise-frontend": "^0.0.34",
-    "@nrwl/schematics": "7.0.2",
+    "@angular-devkit/build-angular": "0.11.4",
+    "@angular-devkit/build-ng-packagr": "0.11.4",
+    "@angular-devkit/core": "7.1.4",
+    "@angular-devkit/schematics": "7.1.4",
+    "@angular/animations": "7.1.4",
+    "@angular/cdk": "7.2.0",
+    "@angular/cli": "7.1.4",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/compiler-cli": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/flex-layout": "7.0.0-beta.22",
+    "@angular/forms": "7.1.4",
+    "@angular/language-service": "7.1.4",
+    "@angular/material": "7.2.0",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/router": "7.1.4",
+    "@graphql-modules/logger": "^0.2.3",
+    "@nrwl/angular-console-enterprise-electron": "file:../nrwl-connect/dist/libs/angular-console-extensions/electron",
+    "@nrwl/angular-console-enterprise-frontend": "file:../nrwl-connect/dist/libs/angular-console-extensions/frontend",
+    "@nrwl/builders": "7.1.1",
+    "@nrwl/schematics": "7.1.1",
+    "@types/electron-store": "^1.3.0",
     "@types/express": "4.16.0",
-    "@types/graphql": "0.13.4",
-    "@types/jasmine": "2.8.12",
-    "@types/jasminewd2": "2.0.6",
+    "@types/get-port": "^4.0.1",
+    "@types/graphql": "14.0.3",
+    "@types/jest": "^23.3.10",
     "@types/node": "8.10.38",
+    "@types/request": "^2.48.1",
+    "@types/semver": "^5.5.0",
+    "@types/shelljs": "^0.8.1",
     "@types/sinon": "5.0.7",
     "@types/sinon-chai": "3.2.1",
+    "@types/strip-json-comments": "^0.0.30",
+    "@types/tmp": "^0.0.33",
     "@types/universal-analytics": "0.4.2",
-    "angular-playground": "5.2.0",
     "apollo-angular": "1.5.0",
     "apollo-angular-link-http": "1.4.0",
     "apollo-cache-inmemory": "1.3.11",
-    "apollo-client": "2.4.7",
     "apollo-link": "1.2.4",
     "apollo-link-error": "1.1.2",
-    "apollo-server-express": "2.2.2",
+    "apollo-server-express": "^2.2.2",
     "codelyzer": "4.5.0",
     "core-js": "2.5.7",
-    "cypress": "3.1.1",
-    "drivelist": "6.4.3",
-    "electron": "2.0.14",
-    "electron-builder": "20.28.4",
+    "cypress": "3.1.4",
+    "dotenv": "6.2.0",
+    "electron": "4.0.0",
+    "electron-builder": "20.38.4",
     "electron-installer-dmg": "2.0.0",
-    "electron-packager": "12.2.0",
+    "electron-packager": "13.0.1",
+    "electron-rebuild": "^1.8.2",
     "electron-store": "2.0.0",
-    "electron-updater": "3.1.2",
+    "electron-updater": "4.0.6",
     "express": "4.16.4",
-    "express-graphql": "0.7.1",
     "fix-path": "2.1.0",
-    "get-port": "3.2.0",
-    "graphql": "0.13.2",
-    "graphql-code-generator": "0.14.5",
-    "graphql-codegen-typescript-apollo-angular": "0.14.5",
-    "graphql-codegen-typescript-common": "0.14.5",
-    "graphql-codegen-typescript-client": "0.14.5",
-    "graphql-codegen-typescript-resolvers": "0.14.5",
-    "graphql-codegen-typescript-server": "0.14.5",
+    "get-port": "4.0.0",
+    "graphql": "^0.13.1",
+    "graphql-code-generator": "0.15.2",
+    "graphql-codegen-typescript-apollo-angular": "0.15.2",
+    "graphql-codegen-typescript-client": "0.15.2",
+    "graphql-codegen-typescript-common": "0.15.2",
+    "graphql-codegen-typescript-resolvers": "0.15.2",
+    "graphql-codegen-typescript-server": "0.15.2",
     "graphql-tag": "2.10.0",
-    "jasmine-core": "3.3.0",
-    "jasmine-marbles": "0.4.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "jest": "23.6.0",
+    "jasmine-marbles": "^0.4.1",
+    "jest": "^23.0.0",
     "jest-jasmine2": "23.6.0",
-    "karma": "3.1.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.4",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.0",
-    "ng-packagr": "4.4.0",
+    "jest-preset-angular": "6.0.1",
+    "ng-packagr": "4.4.5",
     "node-pty-prebuilt": "0.7.6",
     "nps": "5.9.3",
     "nps-utils": "1.7.0",
-    "prettier": "1.15.2",
-    "protractor": "5.4.1",
-    "resolve": "1.8.1",
+    "prettier": "1.15.3",
+    "progress-stream": "2.0.0",
+    "request": "2.88.0",
+    "resolve": "1.9.0",
     "rxjs": "6.3.3",
     "rxjs-compat": "6.3.3",
-    "shelljs": "0.8.2",
+    "shelljs": "0.8.3",
     "strip-json-comments": "2.0.1",
     "ts-jest": "23.10.5",
     "ts-node": "7.0.1",
     "tsickle": "0.34.0",
     "tslib": "1.9.3",
-    "tslint": "5.11.0",
+    "tslint": "5.12.0",
+    "tslint-no-unused-expression-chai": "^0.1.4",
     "typescript": "3.1.6",
-    "universal-analytics": "0.4.20",
-    "uuid": "3.3.2",
-    "xterm": "3.8.0",
-    "zone.js": "0.8.26",
-    "@types/jquery": "3.3.22"
+    "universal-analytics": "^0.4.20",
+    "uuid": "^3.3.2",
+    "vscode": "^1.1.26",
+    "xterm": "3.9.1",
+    "zone.js": "0.8.26"
   },
   "jest": {
     "moduleFileExtensions": [
