@@ -2,19 +2,46 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { CommandOutputComponent } from './command-output.component';
 import { UiModule } from '../ui.module';
 import { CommandStatus, CommandRunner } from '@angular-console/utils';
+import { TerminalFactory } from '../terminal/terminal.factory';
+import { Terminal } from 'xterm';
 
 describe('CommandOutputComponent', () => {
   let component: CommandOutputComponent;
   let fixture: ComponentFixture<CommandOutputComponent>;
-  let mockCommandRunner: jasmine.SpyObj<CommandRunner>;
+  let mockCommandRunner: { stopCommandViaCtrlC: jest.Mock<{}> };
 
   beforeEach(async(() => {
-    mockCommandRunner = jasmine.createSpyObj<CommandRunner>('CommandRunner', [
-      'stopCommandViaCtrlC'
-    ]);
+    window.matchMedia = jest.fn().mockImplementation(query => {
+      return {
+        matches: false,
+        media: query,
+        onchange: null,
+        addListener: jest.fn(),
+        removeListener: jest.fn()
+      };
+    });
+    mockCommandRunner = {
+      stopCommandViaCtrlC: jest.fn()
+    };
     TestBed.configureTestingModule({
       imports: [UiModule],
-      providers: [{ provide: CommandRunner, useValue: mockCommandRunner }]
+      providers: [
+        { provide: CommandRunner, useValue: mockCommandRunner },
+        {
+          provide: TerminalFactory,
+          useValue: {
+            new: () => ({
+              _core: {},
+              open: jest.fn(),
+              write: jest.fn(),
+              rows: NaN,
+              cols: NaN,
+              reset: jest.fn(),
+              resize: jest.fn()
+            })
+          }
+        }
+      ]
     }).compileComponents();
   }));
 
@@ -46,7 +73,7 @@ describe('CommandOutputComponent', () => {
       component.commandResponse.id
     );
 
-    mockCommandRunner.stopCommandViaCtrlC.calls.reset();
+    mockCommandRunner.stopCommandViaCtrlC.mockClear();
     component.commandResponse.status = CommandStatus.TERMINATED;
     document.dispatchEvent(
       new KeyboardEvent('keypress', { ctrlKey: true, key: 'c' })
