@@ -1,5 +1,5 @@
-export const COMMANDS_POLLING = 100;
-export const NODE_JS_INSTALL_POLLING = 100;
+export const COMMANDS_POLLING = 300;
+export const NODE_JS_INSTALL_POLLING = 300;
 export const BASIC_WORKSPACE_POLLING = 5000;
 export const PROJECTS_POLLING = 5000;
 export const TARGET_POLLING = 5000;
